♻️ Rename the subpackage to align with another subpackage naming rule.
